+++ ASCOfficeOdfFile/src/odf/math_elements.cpp
@@ -99,7 +99,7 @@ void math_semantics::add_child_element( xml::sax * Reader, const std::wstring & 
 
 void math_semantics::oox_convert(oox::math_context & Context)
 {
-	for (int i = 0 ; i < content_.size(); i++)
+	for (size_t i = 0 ; i < content_.size(); i++)
     {
 		office_math_element* math_element = dynamic_cast<office_math_element*>(content_[i].get());
         math_element->oox_convert(Context);
