+++ ASCOfficeOdfFileW/source/OdfFormat/odf_conversion_context.cpp
@@ -117,7 +117,7 @@ _mediaitems* odf_conversion_context::mediaitems()
 void odf_conversion_context::end_document()
 {
 	rels	rels_;
-	for (long i=0; i< objects_.size(); i++)
+	for (size_t i = 0; i < objects_.size(); i++)
 	{
 		_object & object = objects_[i];
 		bool isRoot = (i == 0 ? true : false);
